refactor(vulkan): rename classes and files for better consistency
- Replace Vulkan-related class and file names to use "Vulkan" prefix (e.g., VulkanDevice, VulkanAllocator, etc.)
- Remove obsolete files related to the old code structure
- Update all references and includes to reflect the new names

This refactoring improves code clarity and ensures consistency in Vulkan class naming.
