+++ include/vostok/graphics/vulkan/core/vulkan_allocator.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "vostok/graphics/vulkan/core/device.hpp"
-#include "vostok/graphics/vulkan/core/instance.hpp"
-#include "vostok/graphics/vulkan/core/physical_device.hpp"
+#include "vostok/graphics/vulkan/core/vulkan_device.hpp"
+#include "vostok/graphics/vulkan/core/vulkan_instance.hpp"
+#include "vostok/graphics/vulkan/core/vulkan_physical_device.hpp"
 
 #include <expected>
 #include <memory>
@@ -12,26 +12,26 @@
 namespace vostok::graphics::vulkan
 {
 
-class Allocator
+class VulkanAllocator
 {
 public:
     struct CreateInfo
     {
-        Instance *instance = nullptr;
-        PhysicalDevice *physicalDevice = nullptr;
-        Device *device = nullptr;
+        VulkanInstance *instance = nullptr;
+        VulkanPhysicalDevice *physicalDevice = nullptr;
+        VulkanDevice *device = nullptr;
         u32 vulkanApiVersion = VK_API_VERSION_1_4;
     };
 
     static auto create(const CreateInfo &createInfo)
-        -> std::expected<std::unique_ptr<Allocator>, std::string>;
+        -> std::expected<std::unique_ptr<VulkanAllocator>, std::string>;
 
-    ~Allocator();
+    ~VulkanAllocator();
 
-    Allocator(const Allocator &) = delete;
-    auto operator=(const Allocator &) -> Allocator & = delete;
-    Allocator(Allocator &&other) noexcept;
-    auto operator=(Allocator &&other) noexcept -> Allocator &;
+    VulkanAllocator(const VulkanAllocator &) = delete;
+    auto operator=(const VulkanAllocator &) -> VulkanAllocator & = delete;
+    VulkanAllocator(VulkanAllocator &&other) noexcept;
+    auto operator=(VulkanAllocator &&other) noexcept -> VulkanAllocator &;
 
     auto mapMemory(VmaAllocation allocation, void **data)
         -> std::expected<void *, std::string>;
@@ -47,7 +47,7 @@ public:
     [[nodiscard]] auto getHandle() const -> VmaAllocator { return m_allocator; }
 
 private:
-    Allocator(VmaAllocator allocator);
+    VulkanAllocator(VmaAllocator allocator);
 
     VmaAllocator m_allocator = VK_NULL_HANDLE;
 };
