+++ include/vostok/graphics/vulkan/core/vulkan_frame_sync.hpp
@@ -10,43 +10,43 @@
 namespace vostok::graphics::vulkan
 {
 
-struct FrameData
-{
-    VkCommandBuffer commandBuffer = VK_NULL_HANDLE;
-    VkSemaphore imageAvailable = VK_NULL_HANDLE;
-    VkSemaphore renderFinished = VK_NULL_HANDLE;
-    VkFence inFlight = VK_NULL_HANDLE;
-};
+class VulkanDevice;
+class VulkanCommandPool;
 
-class Device;
-class CommandPool;
-
-class FrameSync
+class VulkanFrameSync
 {
 public:
+    struct FrameData
+    {
+        VkCommandBuffer commandBuffer = VK_NULL_HANDLE;
+        VkSemaphore imageAvailable = VK_NULL_HANDLE;
+        VkSemaphore renderFinished = VK_NULL_HANDLE;
+        VkFence inFlight = VK_NULL_HANDLE;
+    };
+
     struct CommandPools
     {
-        CommandPool *graphics = nullptr;
-        CommandPool *transfer = nullptr;
+        VulkanCommandPool *graphics = nullptr;
+        VulkanCommandPool *transfer = nullptr;
     };
 
     struct CreateInfo
     {
-        Device *device = nullptr;
+        VulkanDevice *device = nullptr;
         u32 maxFramesInFlight = 2;
         CommandPools commandPools;
         VkQueue transferQueue = VK_NULL_HANDLE;
     };
 
-    ~FrameSync();
+    ~VulkanFrameSync();
 
-    FrameSync(const FrameSync &) = delete;
-    auto operator=(const FrameSync &) -> FrameSync & = delete;
-    FrameSync(FrameSync &&other) noexcept;
-    auto operator=(FrameSync &&other) noexcept -> FrameSync &;
+    VulkanFrameSync(const VulkanFrameSync &) = delete;
+    auto operator=(const VulkanFrameSync &) -> VulkanFrameSync & = delete;
+    VulkanFrameSync(VulkanFrameSync &&other) noexcept;
+    auto operator=(VulkanFrameSync &&other) noexcept -> VulkanFrameSync &;
 
     static auto create(const CreateInfo &createInfo)
-        -> std::expected<std::unique_ptr<FrameSync>, std::string>;
+        -> std::expected<std::unique_ptr<VulkanFrameSync>, std::string>;
 
     [[nodiscard]] auto getImageAvailableSemaphore() const -> VkSemaphore;
     [[nodiscard]] auto getRenderFinishedSemaphore() const -> VkSemaphore;
@@ -73,29 +73,29 @@ public:
     auto submitTransferCommandBuffer() -> std::expected<void, std::string>;
     auto waitForTransferComplete() -> std::expected<void, std::string>;
 
-    [[nodiscard]] auto getTransferCommandPool() const -> CommandPool *
+    [[nodiscard]] auto getTransferCommandPool() const -> VulkanCommandPool *
     {
         return m_transferCommandPool;
     }
 
 private:
-    FrameSync(
-        Device *device,
+    VulkanFrameSync(
+        VulkanDevice *device,
         const CommandPools &commandPools,
         VkQueue transferQueue,
         u32 maxFramesInFlight
     );
     auto init() -> bool;
 
-    Device *m_device = nullptr;
+    VulkanDevice *m_device = nullptr;
     u32 m_maxFramesInFlight = 2;
     u32 m_currentFrame = 0;
 
     std::vector<FrameData> m_frames;
 
-    CommandPool *m_graphicsCommandPool = nullptr;
+    VulkanCommandPool *m_graphicsCommandPool = nullptr;
 
-    CommandPool *m_transferCommandPool = nullptr;
+    VulkanCommandPool *m_transferCommandPool = nullptr;
     VkQueue m_transferQueue = VK_NULL_HANDLE;
     VkCommandBuffer m_transferCommandBuffer = VK_NULL_HANDLE;
     VkFence m_transferFence = VK_NULL_HANDLE;
