+++ include/vostok/graphics/vulkan/core/vulkan_physical_device.hpp
@@ -1,0 +1,102 @@
+#pragma once
+
+#include "vostok/core/type.hpp"
+
+#include <cassert>
+#include <expected>
+#include <memory>
+#include <optional>
+#include <vector>
+#include <vulkan/vulkan.h>
+#include <vulkan/vulkan_core.h>
+
+namespace vostok::graphics::vulkan
+{
+
+struct QueueFamilyIndices
+{
+    std::optional<u32> graphicsFamily;
+    std::optional<u32> presentFamily;
+    std::optional<u32> computeFamily;
+    std::optional<u32> transferFamily;
+
+    [[nodiscard]] auto isComplete() const -> bool
+    {
+        return graphicsFamily.has_value() && presentFamily.has_value() &&
+               computeFamily.has_value() && transferFamily.has_value();
+    }
+};
+
+class VulkanPhysicalDevice
+{
+public:
+    struct Features
+    {
+        bool geometryShader = false;
+    };
+
+    static auto create(VkInstance instance, VkSurfaceKHR surface)
+        -> std::expected<std::unique_ptr<VulkanPhysicalDevice>, std::string>;
+
+    ~VulkanPhysicalDevice();
+
+    VulkanPhysicalDevice(const VulkanPhysicalDevice &) = delete;
+    auto operator=(const VulkanPhysicalDevice &)
+        -> VulkanPhysicalDevice & = delete;
+    VulkanPhysicalDevice(VulkanPhysicalDevice &&other) noexcept;
+    auto operator=(VulkanPhysicalDevice &&other) noexcept
+        -> VulkanPhysicalDevice &;
+
+    [[nodiscard]] auto getHandle() const -> VkPhysicalDevice
+    {
+        return m_physicalDevice;
+    }
+    [[nodiscard]] auto getProperties() const
+        -> const VkPhysicalDeviceProperties &
+    {
+        return m_properties;
+    }
+    [[nodiscard]] auto getMemoryProperties() const
+        -> const VkPhysicalDeviceMemoryProperties &
+    {
+        return m_memoryProperties;
+    }
+
+    [[nodiscard]] auto getSupportedFeatures() const -> const Features &
+    {
+        return m_supportedFeatures;
+    }
+    [[nodiscard]] auto getQueueFamilyIndices() const
+        -> const QueueFamilyIndices &
+    {
+        return m_queueFamilyIndices;
+    }
+
+    [[nodiscard]] auto isFormatSupported(
+        VkFormat format,
+        VkImageTiling tiling,
+        VkFormatFeatureFlags features
+    ) const -> bool;
+
+    [[nodiscard]] auto checkDeviceExtensionSupport(
+        const std::vector<const char *> &requiredExtensions
+    ) const -> bool;
+
+private:
+    VulkanPhysicalDevice(VkPhysicalDevice physicalDevice, VkSurfaceKHR surface);
+
+    void initializeQueueFamilyIndices(VkSurfaceKHR surface);
+    void queryPhysicalDeviceProperties();
+    void queryPhysicalDeviceFeatures();
+    void determineAvailableExtensions();
+
+    VkPhysicalDevice m_physicalDevice = VK_NULL_HANDLE;
+    VkPhysicalDeviceProperties m_properties = {};
+    VkPhysicalDeviceMemoryProperties m_memoryProperties = {};
+    VkPhysicalDeviceFeatures m_features = {};
+    Features m_supportedFeatures = {};
+    QueueFamilyIndices m_queueFamilyIndices = {};
+    std::vector<VkExtensionProperties> m_supportedExtensions;
+};
+
+} // namespace vostok::graphics::vulkan
